+++ python/ambassador/ir/irambassador.py
@@ -58,6 +58,7 @@ class IRAmbassador (IRResource):
         'regex_type',
         'resolver',
         'error_response_overrides',
+        'header_case_overrides',
         'server_name',
         'service_port',
         'statsd',